enable custom overFetching heuristic and test all files
