+++ packages/runtime-lint/src/rules/overFetching/overFetching.ts
@@ -1,10 +1,18 @@
-import { objectStats } from "../../utils/objectStats.js";
+import { type ObjectWithStats, objectStats } from "../../utils/objectStats.js";
 
 export type OverFetchingConfig = {
 	/**
 	 * Callback to run whenever we detect a json response has been very underused, which could suggest overfetching.
 	 */
 	cb: (url: string) => void;
+	/**
+	 * Heuristic to apply to a json response after a certian time to determine if it has been underused or not. Return true if it is underused.
+	 */
+	// biome-ignore lint/suspicious/noExplicitAny: <explanation>
+	heuristic: <T extends any[] | Record<string, any>>(
+		response: T,
+		responseWithStats: ObjectWithStats<T>,
+	) => boolean;
 };
 
 export const DEFAULT_OVERFETCHING_CONFIG: OverFetchingConfig = {
@@ -12,6 +20,27 @@ export const DEFAULT_OVERFETCHING_CONFIG: OverFetchingConfig = {
 		console.warn(
 			`We detected a json response that was very under-utilized, this might suggest that you are overfetching your api. The url is ${url}`,
 		);
+	},
+	heuristic: (response, responseWithStats) => {
+		if (Array.isArray(response)) {
+			const arrayElemsUnused = Object.values(responseWithStats.__stats).filter(
+				(value) => value.count == null || value.count === 0,
+			).length;
+			if (arrayElemsUnused > response.length / 2) {
+				return true;
+			}
+		} else {
+			// Is a simple object. Only check top-level keys and check if under half of them have been used
+			const numToplevelKeys = Object.keys(responseWithStats).length;
+			const unaccessKeysCount = Object.values(responseWithStats.__stats).filter(
+				(value) => value.count == null || value.count === 0,
+			).length;
+			if (unaccessKeysCount > numToplevelKeys / 2) {
+				return true;
+			}
+		}
+
+		return false;
 	},
 };
 
@@ -24,22 +53,8 @@ export function detectOverfetching<T extends any[] | Record<string, any>>(
 	const statObject = objectStats(response);
 
 	setTimeout(() => {
-		if (Array.isArray(response)) {
-			const arrayElemsUnused = Object.values(statObject.__stats).filter(
-				(value) => value.count == null || value.count === 0,
-			).length;
-			if (arrayElemsUnused > response.length / 2) {
-				config.cb(url);
-			}
-		} else {
-			// Is a simple object. Only check top-level keys and check if under half of them have been used
-			const numToplevelKeys = Object.keys(statObject).length;
-			const unaccessKeysCount = Object.values(statObject.__stats).filter(
-				(value) => value.count == null || value.count === 0,
-			).length;
-			if (unaccessKeysCount > numToplevelKeys / 2) {
-				config.cb(url);
-			}
+		if (config.heuristic(response, statObject)) {
+			config.cb(url);
 		}
 	}, 1000);
